put benchmarker in main

+++ regression-finder/src/main/java/com/items/BenchMark.java
@@ -5,7 +5,7 @@ import com.items.interfaces.CSVItem;
 public class BenchMark implements CSVItem {
 
 	private static final int SECONDS_PER_MINUTE = 60;
-	private static final int MILLISECONDS_PER_SECOND = 1000;
+	private static final double MILLISECONDS_PER_SECOND = 1000;
 	private long time;
 	private String method;
 
@@ -23,7 +23,6 @@ public class BenchMark implements CSVItem {
 		double seconds = time/MILLISECONDS_PER_SECOND;
 		long minutes = (long) Math.floor(seconds/SECONDS_PER_MINUTE);
 		seconds -= minutes*SECONDS_PER_MINUTE;
-
 		return minutes + "," + seconds;
 	}
